+++ services/ADC.py
@@ -1,10 +1,9 @@
 from functools import partial
-from slsdet import Detector, dacIndex, readoutMode, runStatus
 from PyQt5 import QtWidgets
 import pyqtgraph as pg
 
-from bit_utils import bit_is_set, manipulate_bit
-from defines import Defines
+from utils.bit_utils import bit_is_set, manipulate_bit
+from utils.defines import Defines
 
 
 class ADC:
