+++ services/Pattern.py
@@ -2,15 +2,12 @@ import os
 from functools import partial
 from pathlib import Path
 
-from slsdet import Detector, dacIndex, readoutMode, runStatus
 from PyQt5 import QtWidgets
-import pyqtgraph as pg
 import matplotlib.pyplot as plt
 from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
 from matplotlib.backends.backend_qt5agg import NavigationToolbar2QT as NavigationToolbar
-from bit_utils import bit_is_set, manipulate_bit
-from defines import Defines
-from plotPattern import PlotPattern
+from utils.defines import Defines
+from utils.plotPattern import PlotPattern
 
 
 class Pattern:
