+++ services/Signals.py
@@ -2,8 +2,8 @@ from functools import partial
 from PyQt5 import QtWidgets
 import pyqtgraph as pg
 
-from bit_utils import bit_is_set, manipulate_bit
-from defines import Defines
+from utils.bit_utils import bit_is_set, manipulate_bit
+from utils.defines import Defines
 
 
 class Signals:
