+++ services/SlowADCs.py
