+++ utils/decoder.py
@@ -1,4 +1,4 @@
-from defines import *
+from utils.defines import *
 from _decoder import * #bring in the function from the compiled extension
 import numpy as np
